Groq_Chat/requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
python-docx
pymupdf
